Added more onto the fish class
I created more methods for the fish class and added onto existing
methods. The set and get fish methods are removed. I removed the.__gills
attribute in the constructor and replaced it with .__alive instead for
the consume method. There are getters and setters for the speed, status,
swim, position, and size. I changed the get_eaten method to consume and
then I added onto swim_forward as well.